+++ extensions/api/data-management/contractnegotiation/src/test/java/org/eclipse/dataspaceconnector/api/datamanagement/contractnegotiation/service/ContractNegotiationServiceImplTest.java
@@ -28,14 +28,18 @@ import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.Cont
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractOfferRequest;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.command.ContractNegotiationCommand;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractOffer;
+import org.eclipse.dataspaceconnector.sql.translation.EdcQueryException;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentMatcher;
 
 import java.util.UUID;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.eclipse.dataspaceconnector.api.result.ServiceFailure.Reason.CONFLICT;
 import static org.eclipse.dataspaceconnector.api.result.ServiceFailure.Reason.NOT_FOUND;
 import static org.eclipse.dataspaceconnector.spi.response.ResponseStatus.FATAL_ERROR;
@@ -84,6 +88,35 @@ class ContractNegotiationServiceImplTest {
         var result = service.query(QuerySpec.none());
 
         assertThat(result).hasSize(1).first().matches(it -> it.getId().equals("negotiationId"));
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {
+            "contractAgreement.contractStartDate.begin=123455", //invalid path
+            "contractOffers.policy.uid=some-id", //invalid path
+            "contractOffers.policy.assetid=some-id", //wrong case
+            "contractOffers.policy.=some-id", //incomplete path
+    })
+    void query_invalidFilter(String invalidFilter) {
+        var query = QuerySpec.Builder.newInstance()
+                .filter(invalidFilter)
+                .build();
+
+        assertThatThrownBy(() -> service.query(query)).isInstanceOf(EdcQueryException.class);
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {
+            "contractAgreement.contractStartDate=123455",
+            "contractAgreement.assetId=test-asset",
+            "contractAgreement.policy.assignee=123455",
+    })
+    void query_validFilter(String validFilter) {
+        var query = QuerySpec.Builder.newInstance()
+                .filter(validFilter)
+                .build();
+        service.query(query);
+        verify(store).queryNegotiations(query);
     }
 
     @Test