+++ CHANGELOG.md
@@ -61,6 +61,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Fix race condition in `ContractNegotiationIntegrationTest` (#1505)
 * Fix for change in Cosmos DB behavior on missing sort fields (#1514)
 * Effectively removed default LIMIT in SQL Contract Def Store (#1515)
+* Added query validation (#1605)
 
 ## [milestone-4] - 2022-06-07